Permit negative values
According to the GPSD JSON protocol, these values are "numeric", so
should probably be parsed as floats.

https://gpsd.gitlab.io/gpsd/gpsd_json.html

However, the underlying satellite_t data structure defines these values
as (signed) shorts, so an i16 type would be more appropriate.

https://gitlab.com/gpsd/gpsd/blob/release-3.18.1/gps.h#L1811-1850

Note that for elevation, azimuth, and SNR, -1 is used as a sentinal N/A
value.